# blas/gonum: use atomics to distribute work