Optimizes use of a channel for distributing work to workers.

```
$ benchstat before.txt after.txt

name                  old time/op  new time/op  delta
DgemmSmSmSm-32         822ns ± 1%   817ns ± 1%     ~     (p=0.190 n=5+5)
DgemmMedMedMed-32      136µs ± 1%   141µs ± 3%   +3.30%  (p=0.008 n=5+5)
DgemmMedLgMed-32       463µs ± 0%   394µs ± 0%  -14.83%  (p=0.008 n=5+5)
DgemmLgLgLg-32        25.0ms ± 0%  24.9ms ± 0%   -0.51%  (p=0.008 n=5+5)
DgemmLgSmLg-32         697µs ± 0%   673µs ± 1%   -3.49%  (p=0.008 n=5+5)
DgemmLgLgSm-32         811µs ± 0%   735µs ± 0%   -9.36%  (p=0.008 n=5+5)
DgemmHgHgSm-32        72.5ms ± 0%  68.3ms ± 0%   -5.79%  (p=0.008 n=5+5)
DgemmMedMedMedTNT-32   299µs ± 3%   410µs ±16%  +36.95%  (p=0.008 n=5+5)
DgemmMedMedMedNTT-32   145µs ± 3%   147µs ± 3%     ~     (p=0.421 n=5+5)
DgemmMedMedMedTT-32    522µs ±12%   770µs ±20%  +47.43%  (p=0.008 n=5+5)
```

`DgemmMedMedMedTNT` is quite noisy in the measurements because the go scheduler decisions can affect the benchmark results quite significantly here.

If a worker finishes early then it might grab a work item and starve another worker. In MedMedMed case I got a case where the work distribution looked like `[0 1 0 3]`. Channels do avoid this problem due to having some fairness guarantees, however this also means that there's more machinery involved.