@egonelbre 2020-03-13
I suspect that the blockSize is too small. Effectively with that case it one of the workers manages to finish work quick, which ends up causing another one not to get any work at all.

e.g instead of W1=[a], W2=[b], you get W1=[a,b] and W2=[], where `a` and `b` are some amount of work. If we speed up the inner blockSize working, then it is more likely to happen.