@egonelbre 2020-03-13
Inidividual mesurements before:

```
BenchmarkDgemmMedMedMedTNT-32              22246            291153 ns/op
BenchmarkDgemmMedMedMedTNT-32              19556            302772 ns/op
BenchmarkDgemmMedMedMedTNT-32              18391            289925 ns/op
BenchmarkDgemmMedMedMedTNT-32              19795            309674 ns/op
BenchmarkDgemmMedMedMedTNT-32              21306            303764 ns/op
BenchmarkDgemmMedMedMedNTT-32              39760            148944 ns/op
BenchmarkDgemmMedMedMedNTT-32              40512            141514 ns/op
BenchmarkDgemmMedMedMedNTT-32              41151            142475 ns/op
BenchmarkDgemmMedMedMedNTT-32              41095            146271 ns/op
BenchmarkDgemmMedMedMedNTT-32              39919            146396 ns/op
BenchmarkDgemmMedMedMedTT-32               15301            582380 ns/op
BenchmarkDgemmMedMedMedTT-32               13694            458011 ns/op
BenchmarkDgemmMedMedMedTT-32               14316            472757 ns/op
BenchmarkDgemmMedMedMedTT-32               10000            563800 ns/op
BenchmarkDgemmMedMedMedTT-32               10000            535000 ns/op
```

Individual measurements after:
```
BenchmarkDgemmMedMedMedTNT-32              16450            439453 ns/op
BenchmarkDgemmMedMedMedTNT-32              14090            414975 ns/op
BenchmarkDgemmMedMedMedTNT-32              17035            446023 ns/op
BenchmarkDgemmMedMedMedTNT-32              20782            404726 ns/op
BenchmarkDgemmMedMedMedTNT-32              14661            345406 ns/op
BenchmarkDgemmMedMedMedNTT-32              40650            149373 ns/op
BenchmarkDgemmMedMedMedNTT-32              41011            148448 ns/op
BenchmarkDgemmMedMedMedNTT-32              40348            145534 ns/op
BenchmarkDgemmMedMedMedNTT-32              39214            148620 ns/op
BenchmarkDgemmMedMedMedNTT-32              41011            143108 ns/op
BenchmarkDgemmMedMedMedTT-32               10000            689600 ns/op
BenchmarkDgemmMedMedMedTT-32               10184            912510 ns/op
BenchmarkDgemmMedMedMedTT-32               10000            922202 ns/op
BenchmarkDgemmMedMedMedTT-32               10000            692100 ns/op
BenchmarkDgemmMedMedMedTT-32               10000            634301 ns/op
```