@egonelbre 2020-03-13
Indeed, it seems sometimes we get a work distribution like:
```
[0 1 0 3]
```

Channels do have better fairness than a atomic work queue.