@egonelbre 2020-03-13
Anyways, I'm not convinced how representative my benchmarking is. It seems that `DgemmMedMedMed` benchmarks is one the areas where "go scheduler" decisions significantly affect the benchmark result.

Basically, I'm not sure whether this change is an overall improvement or not.