@egonelbre 2020-03-13
Closing in favor of https://github.com/gonum/gonum/pull/1286